remove form options, rename files, update font
